% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANN-package.R
\docType{package}
\name{RANN-package}
\alias{RANN-package}
\alias{RANN}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library
}
\seealso{
\code{\link{nn2}}
}
\keyword{package}
