\docType{package}
\name{RANN-package}
\alias{RANN}
\alias{RANN-package}
\title{Wrapper for Arya and Mount's Approximat Nearest Neighbours (ANN) C++ library}
\description{
  Wrapper for Arya and Mount's Approximat Nearest
  Neighbours (ANN) C++ library
}
\seealso{
  \code{\link{nn}, \link{nn2}}
}
\keyword{package}

