% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RANN-package}
\alias{RANN}
\alias{RANN-package}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library
}
\seealso{
\code{\link{nn2}}
}
\keyword{package}

